requirements.txt
torch


cupy-cuda12x
matplotlib
unittest